updatessss

+++ addtocart.php
@@ -42,7 +42,7 @@ if ($_SERVER['REQUEST_METHOD'] === 'POST') {
     }
 }
 
-if ($_SERVER['REQUEST_METHOD'] === 'DELETE') {
+if ($_SERVER['REQUEST_METHOD'] === 'DELETE' && isset($_GET['id'])) {
     $id = $_GET['id'];
 
     $sql = "DELETE FROM addtocart WHERE id = ?";
@@ -53,10 +53,11 @@ if ($_SERVER['REQUEST_METHOD'] === 'DELETE') {
     if ($stmt->execute()) {
         // Send the deleted product details as JSON response
         http_response_code(200); // OK
-        echo json_encode($deletedProduct);
+        echo json_encode(['status' => 'success', 'message' => 'Add to cart deleted']);
     } else {
         // The delete operation failed
         http_response_code(500); // Internal Server Error
         echo json_encode(['error' => 'Unable to delete cart']);
     }
 }
+
